+++ design-system/src/main/java/com/danbam/design_system/component/Item.kt
@@ -106,7 +106,7 @@ fun FundingItem(
 @Composable
 fun MovieItem(
     item: MovieEntity,
-    onClickItem: (Int) -> Unit,
+    onClickItem: (Long) -> Unit,
 ) {
     ImageButton(
         modifier = Modifier
