+++ domain/src/main/java/com/danbam/domain/entity/MoviePeopleEntity.kt
@@ -1,7 +1,7 @@
 package com.danbam.domain.entity
 
 data class MoviePeopleEntity(
-    val idx: Int,
+    val idx: Long,
     val name: String,
     val profileUrl: String,
 )
